Move pydantic fields in `server/parsec/types.py`
